change20

+++ backend/models/Champs.js
@@ -8,8 +8,8 @@ const Champ = sequelize.define('table_champs', {
     nom: { type: DataTypes.STRING, allowNull: false },
     surface: { type: DataTypes.DECIMAL, allowNull: false },
     type_culture: { type: DataTypes.STRING, allowNull: false },	
-    longitude: { type: DataTypes.DECIMAL, allowNull: true},
-    latitude: { type: DataTypes.DECIMAL, allowNull: true},
+    longitude: { type: DataTypes.FLOAT, allowNull: true},
+    latitude: { type: DataTypes.FLOAT, allowNull: true},
     etat: {type: DataTypes.STRING, allowNull: false }
 }, {
     timestamp: true
